// fixrender/src/types.ts
export interface Supplier {
  id: string;
  name: string;
  phone: string;
  address?: string;
  email?: string;
  rating?: number;
  products?: { productId: string; purchasePrice?: number }[];
  typeOfMerchandise?: string; // Mantener por compatibilidad si es necesario
}